refactor: separate horizontal infinite scroll layout logic from index page code
