+++ Main/html/index.html
@@ -23,8 +23,6 @@
     </nav>
     <div id="mainContainer" class="main-container">
         <div class="post-board-columns">
-            <article class="column main" data-key="home">
-            </article>
         </div class="post-board-columns">
     </div>
 </body>
